+++ CHANGELOG.md
@@ -6,10 +6,12 @@ Todos los cambios notables de este proyecto se documentarán en este archivo.
   > Ver TODO.md
 
 ---
-* ## [v1.0.0] - 2022-05-09
+* ## [v1.0.1] - 2022-05-11
   > Nuevas funcionalidades.
 
   * #### Añadido:
+    - Colores en los resultados.
+    - Función para obtener parámetros.
 
   * #### Cambios:
     - Mejora en la función de comprobación de existencia del archivo .env.